refactor(frontend): split left sidebar into nav/desktop/mobile sections
548-line shell sidebar becomes a 212-line composition; markup verified
byte-identical (className/aria/role multiset diff) across the move.
